using fewer input args via opts struct
