+++ beacon-chain/slasher/detect_attestations_test.go
@@ -2,9 +2,9 @@ package slasher
 
 import (
 	"context"
-	"reflect"
 	"testing"
 
+	types "github.com/prysmaticlabs/eth2-types"
 	dbtest "github.com/prysmaticlabs/prysm/beacon-chain/db/testing"
 	slashertypes "github.com/prysmaticlabs/prysm/beacon-chain/slasher/types"
 	"github.com/prysmaticlabs/prysm/shared/testutil/assert"
@@ -12,173 +12,26 @@ import (
 	logTest "github.com/sirupsen/logrus/hooks/test"
 )
 
-func TestService_groupByValidatorChunkIndex(t *testing.T) {
-	tests := []struct {
-		name   string
-		params *Parameters
-		atts   []*slashertypes.CompactAttestation
-		want   map[uint64][]*slashertypes.CompactAttestation
-	}{
-		{
-			name:   "No attestations returns empty map",
-			params: DefaultParams(),
-			atts:   make([]*slashertypes.CompactAttestation, 0),
-			want:   make(map[uint64][]*slashertypes.CompactAttestation),
-		},
-		{
-			name: "Groups multiple attestations belonging to single validator chunk",
-			params: &Parameters{
-				validatorChunkSize: 2,
-			},
-			atts: []*slashertypes.CompactAttestation{
-				{
-					AttestingIndices: []uint64{0, 1},
-				},
-				{
-					AttestingIndices: []uint64{0, 1},
-				},
-			},
-			want: map[uint64][]*slashertypes.CompactAttestation{
-				0: {
-					{
-						AttestingIndices: []uint64{0, 1},
-					},
-					{
-						AttestingIndices: []uint64{0, 1},
-					},
-				},
-			},
-		},
-		{
-			name: "Groups single attestation belonging to multiple validator chunk",
-			params: &Parameters{
-				validatorChunkSize: 2,
-			},
-			atts: []*slashertypes.CompactAttestation{
-				{
-					AttestingIndices: []uint64{0, 2, 4},
-				},
-			},
-			want: map[uint64][]*slashertypes.CompactAttestation{
-				0: {
-					{
-						AttestingIndices: []uint64{0, 2, 4},
-					},
-				},
-				1: {
-					{
-						AttestingIndices: []uint64{0, 2, 4},
-					},
-				},
-				2: {
-					{
-						AttestingIndices: []uint64{0, 2, 4},
-					},
-				},
-			},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			s := &Service{
-				params: tt.params,
-			}
-			if got := s.groupByValidatorChunkIndex(tt.atts); !reflect.DeepEqual(got, tt.want) {
-				t.Errorf("groupByValidatorChunkIndex() = %v, want %v", got, tt.want)
-			}
-		})
-	}
-}
-
-func TestService_groupByChunkIndex(t *testing.T) {
-	tests := []struct {
-		name   string
-		params *Parameters
-		atts   []*slashertypes.CompactAttestation
-		want   map[uint64][]*slashertypes.CompactAttestation
-	}{
-		{
-			name:   "No attestations returns empty map",
-			params: DefaultParams(),
-			atts:   make([]*slashertypes.CompactAttestation, 0),
-			want:   make(map[uint64][]*slashertypes.CompactAttestation),
-		},
-		{
-			name: "Groups multiple attestations belonging to single chunk",
-			params: &Parameters{
-				chunkSize:     2,
-				historyLength: 3,
-			},
-			atts: []*slashertypes.CompactAttestation{
-				{
-					Source: 0,
-				},
-				{
-					Source: 1,
-				},
-			},
-			want: map[uint64][]*slashertypes.CompactAttestation{
-				0: {
-					{
-						Source: 0,
-					},
-					{
-						Source: 1,
-					},
-				},
-			},
-		},
-		{
-			name: "Groups multiple attestations belonging to multiple chunks",
-			params: &Parameters{
-				chunkSize:     2,
-				historyLength: 3,
-			},
-			atts: []*slashertypes.CompactAttestation{
-				{
-					Source: 0,
-				},
-				{
-					Source: 1,
-				},
-				{
-					Source: 2,
-				},
-			},
-			want: map[uint64][]*slashertypes.CompactAttestation{
-				0: {
-					{
-						Source: 0,
-					},
-					{
-						Source: 1,
-					},
-				},
-				1: {
-					{
-						Source: 2,
-					},
-				},
-			},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			s := &Service{
-				params: tt.params,
-			}
-			if got := s.groupByChunkIndex(tt.atts); !reflect.DeepEqual(got, tt.want) {
-				t.Errorf("groupByChunkIndex() = %v, want %v", got, tt.want)
-			}
-		})
-	}
-}
-
-func TestService_loadChunk(t *testing.T) {
+func TestService_loadChunk_MinSpan(t *testing.T) {
 	beaconDB := dbtest.SetupDB(t)
 
 	// Check if the chunk at chunk index already exists in-memory.
-	existingChunk := EmptyMinSpanChunksSlice(DefaultParams())
+	kind := slashertypes.MinSpan
+	params := DefaultParams()
+	existingChunk := EmptyMinSpanChunksSlice(params)
+	emptyChunk := EmptyMinSpanChunksSlice(params)
+	validatorIdx := types.ValidatorIndex(0)
+	epochInChunk := types.Epoch(0)
+	targetEpoch := types.Epoch(2)
+	err := setChunkDataAtEpoch(
+		params,
+		existingChunk.Chunk(),
+		validatorIdx,
+		epochInChunk,
+		targetEpoch,
+	)
+	require.NoError(t, err)
+	require.DeepNotEqual(t, existingChunk, emptyChunk)
 	chunkIdx := uint64(1)
 	updatedChunks := map[uint64]Chunker{
 		chunkIdx: existingChunk,
@@ -189,24 +42,110 @@ func TestService_loadChunk(t *testing.T) {
 			Database: beaconDB,
 		},
 	}
-	received, err := s.loadChunk(updatedChunks, 0, chunkIdx, slashertypes.MinSpan)
+	received, err := s.loadChunk(updatedChunks, 0, chunkIdx, kind)
 	require.NoError(t, err)
 	require.DeepEqual(t, existingChunk, received)
 
 	// If a chunk at a chunk index does not exist, ensure it
 	// is initialized as an empty chunk.
 	chunkIdx = uint64(2)
-	received, err = s.loadChunk(updatedChunks, 0, chunkIdx, slashertypes.MinSpan)
+	received, err = s.loadChunk(updatedChunks, 0, chunkIdx, kind)
+	require.NoError(t, err)
+	require.DeepEqual(t, emptyChunk, received)
+
+	// Save chunks to disk, the load them properly from disk.
+	updatedChunks = map[uint64]Chunker{
+		2: existingChunk,
+		4: existingChunk,
+		6: existingChunk,
+	}
+	err = s.saveUpdatedChunks(
+		context.Background(),
+		kind,
+		updatedChunks,
+		0,
+	)
+	require.NoError(t, err)
+	// Check if the retrieved chunks match what we just saved to disk.
+	for chunkIdx = range updatedChunks {
+		input := make(map[uint64]Chunker)
+		received, err := s.loadChunk(input, 0, chunkIdx, kind)
+		require.NoError(t, err)
+		require.DeepEqual(t, existingChunk, received)
+	}
+}
+
+func TestService_loadChunk_MaxSpan(t *testing.T) {
+	beaconDB := dbtest.SetupDB(t)
+
+	// Check if the chunk at chunk index already exists in-memory.
+	params := DefaultParams()
+	kind := slashertypes.MaxSpan
+	emptyChunk := EmptyMaxSpanChunksSlice(params)
+	existingChunk := EmptyMaxSpanChunksSlice(params)
+	validatorIdx := types.ValidatorIndex(0)
+	epochInChunk := types.Epoch(0)
+	targetEpoch := types.Epoch(2)
+	err := setChunkDataAtEpoch(
+		params,
+		existingChunk.Chunk(),
+		validatorIdx,
+		epochInChunk,
+		targetEpoch,
+	)
+	require.NoError(t, err)
+	require.DeepNotEqual(t, existingChunk, emptyChunk)
+	chunkIdx := uint64(1)
+	updatedChunks := map[uint64]Chunker{
+		chunkIdx: existingChunk,
+	}
+	s := &Service{
+		params: DefaultParams(),
+		serviceCfg: &ServiceConfig{
+			Database: beaconDB,
+		},
+	}
+	received, err := s.loadChunk(updatedChunks, 0, chunkIdx, kind)
 	require.NoError(t, err)
 	require.DeepEqual(t, existingChunk, received)
 
+	// If a chunk at a chunk index does not exist, ensure it
+	// is initialized as an empty chunk.
+	chunkIdx = uint64(2)
+	received, err = s.loadChunk(updatedChunks, 0, chunkIdx, kind)
+	require.NoError(t, err)
+	require.DeepEqual(t, emptyChunk, received)
+
 	// Save chunks to disk, the load them properly from disk.
+	updatedChunks = map[uint64]Chunker{
+		2: existingChunk,
+		4: existingChunk,
+		6: existingChunk,
+	}
+	err = s.saveUpdatedChunks(
+		context.Background(),
+		kind,
+		updatedChunks,
+		0,
+	)
+	require.NoError(t, err)
+	// Check if the retrieved chunks match what we just saved to disk.
+	for chunkIdx = range updatedChunks {
+		input := make(map[uint64]Chunker)
+		received, err := s.loadChunk(input, 0, chunkIdx, kind)
+		require.NoError(t, err)
+		require.DeepEqual(t, existingChunk, received)
+	}
 }
 
 func TestService_processQueuedAttestations(t *testing.T) {
 	hook := logTest.NewGlobal()
+	beaconDB := dbtest.SetupDB(t)
 	s := &Service{
 		params: DefaultParams(),
+		serviceCfg: &ServiceConfig{
+			Database: beaconDB,
+		},
 		attestationQueue: []*slashertypes.CompactAttestation{
 			{
 				AttestingIndices: []uint64{0, 1},
