+++ beacon-chain/slasher/receive_test.go
@@ -112,96 +112,3 @@ func TestSlasher_receiveAttestations_OnlyValidAttestations(t *testing.T) {
 	}
 	require.DeepEqual(t, wanted, s.attestationQueue)
 }
-
-func Test_validateAttestationIntegrity(t *testing.T) {
-	tests := []struct {
-		name string
-		att  *ethpb.IndexedAttestation
-		want bool
-	}{
-		{
-			name: "Nil attestation returns false",
-			att:  nil,
-			want: false,
-		},
-		{
-			name: "Nil attestation data returns false",
-			att:  &ethpb.IndexedAttestation{},
-			want: false,
-		},
-		{
-			name: "Nil attestation source and target returns false",
-			att: &ethpb.IndexedAttestation{
-				Data: &ethpb.AttestationData{},
-			},
-			want: false,
-		},
-		{
-			name: "Nil attestation source and good target returns false",
-			att: &ethpb.IndexedAttestation{
-				Data: &ethpb.AttestationData{
-					Target: &ethpb.Checkpoint{},
-				},
-			},
-			want: false,
-		},
-		{
-			name: "Nil attestation target and good source returns false",
-			att: &ethpb.IndexedAttestation{
-				Data: &ethpb.AttestationData{
-					Source: &ethpb.Checkpoint{},
-				},
-			},
-			want: false,
-		},
-		{
-			name: "Source > target returns false",
-			att: &ethpb.IndexedAttestation{
-				Data: &ethpb.AttestationData{
-					Source: &ethpb.Checkpoint{
-						Epoch: 1,
-					},
-					Target: &ethpb.Checkpoint{
-						Epoch: 0,
-					},
-				},
-			},
-			want: false,
-		},
-		{
-			name: "Source == target returns false",
-			att: &ethpb.IndexedAttestation{
-				Data: &ethpb.AttestationData{
-					Source: &ethpb.Checkpoint{
-						Epoch: 1,
-					},
-					Target: &ethpb.Checkpoint{
-						Epoch: 1,
-					},
-				},
-			},
-			want: false,
-		},
-		{
-			name: "Source < target returns true",
-			att: &ethpb.IndexedAttestation{
-				Data: &ethpb.AttestationData{
-					Source: &ethpb.Checkpoint{
-						Epoch: 1,
-					},
-					Target: &ethpb.Checkpoint{
-						Epoch: 2,
-					},
-				},
-			},
-			want: true,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if got := validateAttestationIntegrity(tt.att); got != tt.want {
-				t.Errorf("validateAttestationIntegrity() = %v, want %v", got, tt.want)
-			}
-		})
-	}
-}
